// Client/src/types/template.type.ts
export type Template = {
  name: string
  serialNo: string
  dateOfRegistration: string
  gender: string
  district: string
  clinicRegNo: string
  address: string
  contactNo: string
  occupation: string
  diagnosis: string
  consultantResponsible: string
  specialNotes: string
  NIC: string
  starPriorityLevel: number
  hover: number
  dateOfBirth: string
  race: string
  religion: string
}